[Feat] 브랜드 카테고리 분리 및 브랜드 로고 이미지 추가
